+++ packages/angular-workspace/spright-angular/chat/conversation/spright-chat-conversation.directive.ts
@@ -1,4 +1,5 @@
-import { Directive } from '@angular/core';
+import { Directive, ElementRef, Input, Renderer2 } from '@angular/core';
+import { toBooleanProperty, type BooleanValueOrAttribute } from '@ni/nimble-angular/internal-utilities';
 import { type ChatConversation, chatConversationTag } from '@ni/spright-components/dist/esm/chat/conversation';
 
 export type { ChatConversation };
@@ -11,4 +12,14 @@ export { chatConversationTag };
     selector: 'spright-chat-conversation',
     standalone: false
 })
-export class SprightChatConversationDirective { }
+export class SprightChatConversationDirective {
+    public get autoScroll(): boolean {
+        return this.elementRef.nativeElement.autoScroll;
+    }
+
+    @Input('auto-scroll') public set autoScroll(value: BooleanValueOrAttribute) {
+        this.renderer.setProperty(this.elementRef.nativeElement, 'autoScroll', toBooleanProperty(value));
+    }
+
+    public constructor(private readonly renderer: Renderer2, private readonly elementRef: ElementRef<ChatConversation>) {}
+}