+++ packages/spright-components/CHANGELOG.md
@@ -1,8 +1,32 @@
 # Change Log - @ni/spright-components
 
-<!-- This log was last generated on Wed, 17 Jun 2026 21:07:03 GMT and should not be manually modified. -->
+<!-- This log was last generated on Sat, 27 Jun 2026 00:25:17 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 6.21.0
+
+Sat, 27 Jun 2026 00:25:17 GMT
+
+### Minor changes
+
+- Add scrolling behavior to chat. ([ni/nimble@fc9dd14](https://github.com/ni/nimble/commit/fc9dd14cc8ab4d541f7a368154003dea5b576105))
+
+## 6.20.4
+
+Thu, 25 Jun 2026 20:51:07 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v35.11.0
+
+## 6.20.3
+
+Fri, 19 Jun 2026 17:31:59 GMT
+
+### Patches
+
+- Fix send button not being re-enabled ([ni/nimble@2e05c1e](https://github.com/ni/nimble/commit/2e05c1e8606452bb54181622b24536c639a63627))
 
 ## 6.20.2